requirements.txt
aioquic==1.0.0
anyio==4.2.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.2.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
Brotli==1.1.0
build==1.2.2.post1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.3
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
executing==1.2.0
fastjsonschema==2.19.1
filelock==3.13.1
firebase-admin==6.5.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.49.0
fqdn==1.5.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
h2==4.1.0
hererocks==0.25.1
hpack==4.0.0
httplib2==0.22.0
hyperframe==6.0.1
idna==3.6
iopath==0.1.10
ipykernel==6.23.2
ipython==8.14.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.2.0
jupyter_core==5.3.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kaitaistruct==0.10
kiwisolver==1.4.5
ldap3==2.9.1
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mistune==3.0.2
mitmproxy==10.3.0
mitmproxy-macos==0.5.2
mitmproxy_rs==0.5.2
msgpack==1.0.8
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.6
notebook==7.0.7
notebook_shim==0.2.3
numpy==2.0.0
opencv-python==4.9.0.80
overrides==7.7.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
pip-tools==7.4.1
platformdirs==4.2.0
portalocker==2.8.2
prometheus-client==0.19.0
prompt-toolkit==3.0.38
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.5
ptyprocess==0.7.0
publicsuffix2==2.20191221
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic_core==2.18.4
Pygments==2.15.1
PyJWT==2.8.0
pylsqpack==0.3.18
pynvim==0.5.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdfium2==4.28.0
pyperclip==1.8.2
pyproject_hooks==1.2.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
service-identity==24.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.6.2
terminado==0.18.0
tinycss2==1.2.1
tornado==6.3.2
tqdm==4.66.2
traitlets==5.9.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
urwid-mitmproxy==2.1.2.1
virtualenv==20.25.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
widgetsnbextension==4.0.9
wsproto==1.2.0
zstandard==0.22.0
